requirements.txt
certifi==2019.3.9
chardet==3.0.4
configobj==4.7.2
decorator==3.4.0
Django==1.11.22
django-cors-headers==3.0.2
djangorestframework==3.9.4
futures==3.2.0
grpc==0.3.post19
idna==2.8
iniparse==0.4
IPy==0.75
kitchen==1.1.1
msgpack-python==0.5.6
perf==0.1
policycoreutils-default-encoding==0.1
prometheus-client==0.6.0
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
python-consul==1.1.0
python-linux-procfs==0.4.9
pytz==2019.1
pyudev==0.15
pyxattr==0.5.1
redis==3.3.2
requests==2.22.0
schedutils==0.4
seobject==0.1
sepolicy==1.1
six==1.12.0
slip==0.4.0
slip.dbus==0.4.0
urlgrabber==3.10
urllib3==1.25.3
yum-metadata-parser==1.1.4
